pyproject.toml
[project]
name = "python-openscad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anchorscad-core>=0.2.4",
    "numpy>=2.3.5",
    "pythonopenscad>=2.2.19",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/keystone"]
